requirements.txt
pymongo>=4.6,<5
python-dotenv>=1.0,<2
pyarrow>=17.0,<18
fastavro>=1.9,<2
google-cloud-storage>=2.16,<3 